// src/ui/generated-types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Date: any;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
};
// ...
export type Administrator = Node & {
  __typename?: 'Administrator';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  firstName: Scalars['String'];
  lastName: Scalars['String'];
  emailAddress: Scalars['String'];
  user: User;
  customFields?: Maybe<Scalars['JSON']>;
};

export type AdministratorFilterParameter = {
  id?: Maybe<IdOperators>;
  createdAt?: Maybe<DateOperators>;
  updatedAt?: Maybe<DateOperators>;
  firstName?: Maybe<StringOperators>;
  lastName?: Maybe<StringOperators>;
  emailAddress?: Maybe<StringOperators>;
};

export type AdministratorList = PaginatedList & {
  __typename?: 'AdministratorList';
  items: Array<Administrator>;
  totalItems: Scalars['Int'];
};
// ...
export type Asset = Node & {
  __typename?: 'Asset';
  tags: Array<Tag>;
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  name: Scalars['String'];
  type: AssetType;
  fileSize: Scalars['Int'];
  mimeType: Scalars['String'];
  width: Scalars['Int'];
  height: Scalars['Int'];
  source: Scalars['String'];
  preview: Scalars['String'];
  focalPoint?: Maybe<Coordinate>;
  customFields?: Maybe<Scalars['JSON']>;
};
// ...
export type AssetList = PaginatedList & {
  __typename?: 'AssetList';
  items: Array<Asset>;
  totalItems: Scalars['Int'];
};
// ...
export type Channel = Node & {
  __typename?: 'Channel';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  code: Scalars['String'];
  token: Scalars['String'];
  defaultTaxZone?: Maybe<Zone>;
  defaultShippingZone?: Maybe<Zone>;
  defaultLanguageCode: LanguageCode;
  currencyCode: CurrencyCode;
  pricesIncludeTax: Scalars['Boolean'];
  customFields?: Maybe<Scalars['JSON']>;
};

/**
 * Returned when the default LanguageCode of a Channel is no longer found in the `availableLanguages`
 * of the GlobalSettings
 */
export type ChannelDefaultLanguageError = ErrorResult & {
  __typename?: 'ChannelDefaultLanguageError';
  errorCode: ErrorCode;
  message: Scalars['String'];
  language: Scalars['String'];
  channelCode: Scalars['String'];
};

export type Collection = Node & {
  __typename?: 'Collection';
  isPrivate: Scalars['Boolean'];
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  languageCode?: Maybe<LanguageCode>;
  name: Scalars['String'];
  slug: Scalars['String'];
  breadcrumbs: Array<CollectionBreadcrumb>;
  position: Scalars['Int'];
  description: Scalars['String'];
  featuredAsset?: Maybe<Asset>;
  assets: Array<Asset>;
  parent?: Maybe<Collection>;
  children?: Maybe<Array<Collection>>;
  filters: Array<ConfigurableOperation>;
  translations: Array<CollectionTranslation>;
  productVariants: ProductVariantList;
  customFields?: Maybe<Scalars['JSON']>;
};
// ...
export type CollectionList = PaginatedList & {
  __typename?: 'CollectionList';
  items: Array<Collection>;
  totalItems: Scalars['Int'];
};
// ...
export type ConfigurableOperationDefinition = {
  __typename?: 'ConfigurableOperationDefinition';
  code: Scalars['String'];
  args: Array<ConfigArgDefinition>;
  description: Scalars['String'];
};
// ...
export type CountryList = PaginatedList & {
  __typename?: 'CountryList';
  items: Array<Country>;
  totalItems: Scalars['Int'];
};
// ...
export type CreateNfeSettingShippingMethodInput = {
  data: NfeSettingShippingMethodInput;
  methodId: Scalars['ID'];
};
// ...
export type CurrentUser = {
  __typename?: 'CurrentUser';
  id: Scalars['ID'];
  identifier: Scalars['String'];
  channels: Array<CurrentUserChannel>;
};
// ...
export type CustomerGroupList = PaginatedList & {
  __typename?: 'CustomerGroupList';
  items: Array<CustomerGroup>;
  totalItems: Scalars['Int'];
};
// ...
export type CustomerList = PaginatedList & {
  __typename?: 'CustomerList';
  items: Array<Customer>;
  totalItems: Scalars['Int'];
};
// ...
export type NfeSetting = Node & {
  __typename?: 'NfeSetting';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  type: NfeSettingType;
  data: Scalars['JSON'];
  methodId?: Maybe<Scalars['ID']>;
};
// ...
export type NfeSettingShippingMethodInput = {
  cnpj?: Maybe<Scalars['String']>;
  razao_social?: Maybe<Scalars['String']>;
  ie?: Maybe<Scalars['String']>;
  cpf?: Maybe<Scalars['String']>;
  nome_completo?: Maybe<Scalars['String']>;
  endereco?: Maybe<Scalars['String']>;
  uf?: Maybe<Scalars['String']>;
  cidade?: Maybe<Scalars['String']>;
  cep?: Maybe<Scalars['String']>;
  placa?: Maybe<Scalars['String']>;
  uf_veiculo?: Maybe<Scalars['String']>;
  rntc?: Maybe<Scalars['String']>;
  reboque?: Maybe<NfeSettingShippingMethodReboqueInput>;
};

export type NfeSettingShippingMethodReboque = {
  __typename?: 'NfeSettingShippingMethodReboque';
  placa?: Maybe<Scalars['String']>;
  uf_veiculo?: Maybe<Scalars['String']>;
  rntc?: Maybe<Scalars['String']>;
  vagao?: Maybe<Scalars['String']>;
  balsa?: Maybe<Scalars['String']>;
};

export type NfeSettingShippingMethodReboqueInput = {
  placa?: Maybe<Scalars['String']>;
  uf_veiculo?: Maybe<Scalars['String']>;
  rntc?: Maybe<Scalars['String']>;
  vagao?: Maybe<Scalars['String']>;
  balsa?: Maybe<Scalars['String']>;
};

export type NfeSettingSortParameter = {
  id?: Maybe<SortOrder>;
  createdAt?: Maybe<SortOrder>;
  updatedAt?: Maybe<SortOrder>;
  methodId?: Maybe<SortOrder>;
};

export enum NfeSettingType {
  PaymentMethod = 'PaymentMethod',
  ShippingMethod = 'ShippingMethod',
  Custom = 'Custom'
}
// ...
export type Node = {
  id: Scalars['ID'];
};
// ...
export type PaymentMethod = Node & {
  __typename?: 'PaymentMethod';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  name: Scalars['String'];
  code: Scalars['String'];
  description: Scalars['String'];
  enabled: Scalars['Boolean'];
  checker?: Maybe<ConfigurableOperation>;
  handler: ConfigurableOperation;
  customFields?: Maybe<Scalars['JSON']>;
};
// ...
export type ProductVariantList = PaginatedList & {
  __typename?: 'ProductVariantList';
  items: Array<ProductVariant>;
  totalItems: Scalars['Int'];
};
// ...
export type Query = {
  __typename?: 'Query';
  administrators: AdministratorList;
  administrator?: Maybe<Administrator>;
  activeAdministrator?: Maybe<Administrator>;
  /** Get a list of Assets */
  assets: AssetList;
  /** Get a single Asset by id */
  asset?: Maybe<Asset>;
  me?: Maybe<CurrentUser>;
  channels: Array<Channel>;
  channel?: Maybe<Channel>;
  activeChannel: Channel;
  collections: CollectionList;
  /** Get a Collection either by id or slug. If neither id nor slug is specified, an error will result. */
  collection?: Maybe<Collection>;
  collectionFilters: Array<ConfigurableOperationDefinition>;
  /** Used for real-time previews of the contents of a Collection */
  previewCollectionVariants: ProductVariantList;
  countries: CountryList;
  country?: Maybe<Country>;
  customerGroups: CustomerGroupList;
  customerGroup?: Maybe<CustomerGroup>;
  customers: CustomerList;
  customer?: Maybe<Customer>;
  facets: FacetList;
  facet?: Maybe<Facet>;
  facetValues: FacetValueList;
  globalSettings: GlobalSettings;
  job?: Maybe<Job>;
  jobs: JobList;
  jobsById: Array<Job>;
  jobQueues: Array<JobQueue>;
  jobBufferSize: Array<JobBufferSize>;
  order?: Maybe<Order>;
  orders: OrderList;
  /** Returns a list of eligible shipping methods for the draft Order */
  eligibleShippingMethodsForDraftOrder: Array<ShippingMethodQuote>;
  paymentMethods: PaymentMethodList;
  paymentMethod?: Maybe<PaymentMethod>;
  paymentMethodEligibilityCheckers: Array<ConfigurableOperationDefinition>;
  paymentMethodHandlers: Array<ConfigurableOperationDefinition>;
  productOptionGroups: Array<ProductOptionGroup>;
  productOptionGroup?: Maybe<ProductOptionGroup>;
  search: SearchResponse;
  pendingSearchIndexUpdates: Scalars['Int'];
  /** List Products */
  products: ProductList;
  /** Get a Product either by id or slug. If neither id nor slug is specified, an error will result. */
  product?: Maybe<Product>;
  /** List ProductVariants either all or for the specific product. */
  productVariants: ProductVariantList;
  /** Get a ProductVariant by id */
  productVariant?: Maybe<ProductVariant>;
  promotion?: Maybe<Promotion>;
  promotions: PromotionList;
  promotionConditions: Array<ConfigurableOperationDefinition>;
  promotionActions: Array<ConfigurableOperationDefinition>;
  roles: RoleList;
  role?: Maybe<Role>;
  shippingMethods: ShippingMethodList;
  shippingMethod?: Maybe<ShippingMethod>;
  shippingEligibilityCheckers: Array<ConfigurableOperationDefinition>;
  shippingCalculators: Array<ConfigurableOperationDefinition>;
  fulfillmentHandlers: Array<ConfigurableOperationDefinition>;
  testShippingMethod: TestShippingMethodResult;
  testEligibleShippingMethods: Array<ShippingMethodQuote>;
  tag: Tag;
  tags: TagList;
  taxCategories: Array<TaxCategory>;
  taxCategory?: Maybe<TaxCategory>;
  taxRates: TaxRateList;
  taxRate?: Maybe<TaxRate>;
  zones: Array<Zone>;
  zone?: Maybe<Zone>;
  orderNfe?: Maybe<Nfe>;
  nfeCompany?: Maybe<NfeCompany>;
  nfeCompanyByChannel?: Maybe<NfeCompany>;
  nfeCompanies: NfeCompanyList;
  nfeSetting?: Maybe<NfeSetting>;
  nfeSettings: NfeSettingList;
  nfeSettingByTypeAndMethodId?: Maybe<NfeSetting>;
  platform?: Maybe<Platform>;
  platforms: PlatformList;
  platformResource?: Maybe<PlatformResource>;
  platformResources: PlatformResourceList;
  webhook?: Maybe<Scalars['String']>;
  packages: PackageList;
  package: Package;
};
// ...
export type ShippingMethod = Node & {
  __typename?: 'ShippingMethod';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  languageCode: LanguageCode;
  code: Scalars['String'];
  name: Scalars['String'];
  description: Scalars['String'];
  fulfillmentHandlerCode: Scalars['String'];
  checker: ConfigurableOperation;
  calculator: ConfigurableOperation;
  translations: Array<ShippingMethodTranslation>;
  customFields?: Maybe<Scalars['JSON']>;
};
// ...
export type UpdateNfeSettingShippingMethodInput = {
  id: Scalars['ID'];
  data: NfeSettingShippingMethodInput;
};
// ...
  export type Fragment = NfeListItemFragment;
// ...
export namespace NfeSetting {
  export type Fragment = NfeSettingFragment;
}
// ...
  export type Items = NonNullable<(NonNullable<(NonNullable<NfeProductFragment['orderLine']>)['items']>)[number]>;
// ...
  export type Variables = CreateNfeCompanyMutationVariables;
// ...
export namespace GetNfeSetting {
  export type Variables = GetNfeSettingQueryVariables;
  export type Query = GetNfeSettingQuery;
  export type NfeSetting = (NonNullable<GetNfeSettingQuery['nfeSetting']>);
}

export namespace GetNfeSettings {
  export type Variables = GetNfeSettingsQueryVariables;
  export type Query = GetNfeSettingsQuery;
  export type NfeSettings = (NonNullable<GetNfeSettingsQuery['nfeSettings']>);
  export type Items = NonNullable<(NonNullable<(NonNullable<GetNfeSettingsQuery['nfeSettings']>)['items']>)[number]>;
}
// ...
export type NfeSettingFragment = (
  { __typename?: 'NfeSetting' }
  & Pick<NfeSetting, 'id' | 'createdAt' | 'updatedAt' | 'type' | 'data' | 'methodId'>
);
// ...
export type CreateNfeCompanyMutationVariables = Exact<{
  input: CreateNfeCompanyInput;
}>;
// ...
export type GetNfeSettingQueryVariables = Exact<{
  id: Scalars['ID'];
}>;


export type GetNfeSettingQuery = (
  { __typename?: 'Query' }
  & { nfeSetting?: Maybe<(
    { __typename?: 'NfeSetting' }
    & NfeSettingFragment
  )> }
);

export type GetNfeSettingsQueryVariables = Exact<{
  options?: Maybe<NfeSettingListOptions>;
}>;


export type GetNfeSettingsQuery = (
  { __typename?: 'Query' }
  & { nfeSettings: (
    { __typename?: 'NfeSettingList' }
    & Pick<NfeSettingList, 'totalItems'>
    & { items: Array<(
      { __typename?: 'NfeSetting' }
      & NfeSettingFragment
    )> }
  ) }
);
// ...
export type CreateNfeSettingShippingMethodMutationVariables = Exact<{
  input: CreateNfeSettingShippingMethodInput;
}>;


export type CreateNfeSettingShippingMethodMutation = (
  { __typename?: 'Mutation' }
  & { createNfeSettingShippingMethod: (
    { __typename?: 'NfeSetting' }
    & NfeSettingFragment
  ) }
);

export type UpdateNfeSettingShippingMethodMutationVariables = Exact<{
  input: UpdateNfeSettingShippingMethodInput;
}>;


export type UpdateNfeSettingShippingMethodMutation = (
  { __typename?: 'Mutation' }
  & { updateNfeSettingShippingMethod: (
    { __typename?: 'NfeSetting' }
    & NfeSettingFragment
  ) }
);